requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
plotly>=5.18.0
halo>=0.0.31
pytest>=7.4.0
responses>=0.24.0
python-slugify>=8.0.0
ruff>=0.1.0 
PyYAML 